backend/requirements.txt
django
djangorestframework
django-environ
django-cors-headers

torch==1.8.1+cu111
protobuf
sentencepiece
transformers